requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
redis==5.1.0
python-dotenv==1.0.1
pydantic-settings==2.5.2
alembic==1.13.3
